Split profile module
