+++ crates/engine/project/src/profile/mod.rs
@@ -1,0 +1,9 @@
+mod memory;
+mod metrics;
+
+pub(crate) use self::memory::BuildMemorySampler;
+pub(crate) use self::metrics::{metric, profile_descriptors, record_build_checkpoint};
+pub use self::{
+    memory::{BuildProcessMemory, ProcessMemorySampler},
+    metrics::BUILD_CHECKPOINTS,
+};
